Fixed columns in migrations and Added Seeders with Factories

+++ database/migrations/2019_11_09_115955_create_coefficients_table.php
@@ -19,7 +19,7 @@ class CreateCoefficientsTable extends Migration
             $table->unsignedInteger('commerce_value_id');
 
             $table->unsignedInteger('amount');
-            $table->unsignedInteger('percent');
+            $table->float('percent');
 
             $table->timestamps();
